+++ server/src/main/java/backend/dao/ISpec.java
@@ -33,6 +33,6 @@ public interface ISpec extends JpaRepository<Spec,Long> {
     @Transactional
     @Query(value = "INSERT INTO spec (spec_name, group_id, spec_detail, spec_value)\n" +
             "VALUES (:specName, :groupId, :specDetail, :specValue)", nativeQuery = true)
-    void insertSpec(@Param("specName") String specName, @Param("groupId") int groupId,
+    void insertSpec(@Param("specName") String specName, @Param("groupId") Long groupId,
                        @Param("specDetail") String specDetail, @Param("specValue") String specValue);
 }
